Blog/requirements.txt
streamlit>=1.38.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
python-dotenv>=1.0.1
